+++ Practica1-OHNO/Engine/src/main/java/es/ucm/gdv/blas/oses/carreau/lib/Engine/Interfaces/Font.java
@@ -3,6 +3,6 @@ package es.ucm.gdv.blas.oses.carreau.lib.Engine.Interfaces;
 public interface Font {
 
     String getFontName();
-    int getFontSize();
+    float getFontSize();
     boolean isBold();
 }
